requirements.txt
Django>=1.5,<1.5.99 
Embedly==0.4.3
South==0.7.6
boto==2.5.2
dj-database-url==0.2.1
django-heroku-memcacheify==0.4
django-pylibmc-sasl==0.2.4
django-storages==1.1.5
django-uuidfield==0.2
gevent==0.13.8
gunicorn==0.14.6
mailsnake==1.6.1
psycopg2==2.4.1
pusher==0.6
pycrypto==2.5
pylibmc==1.2.3
python-dateutil==2.1
python-memcached==1.53
pytz==2012j
raven==3.1.17
redis==2.7.2
requests==0.12.1
selenium==2.30.0
ssh==1.7.13
uuid==1.30
virtualenv==1.7.1.2
wsgiref==0.1.2

# just so I have it on heroku:
ipython 
